# cms-server/docs/swagger.yaml
basePath: /
definitions:
  business.BusinessExample:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      ID:
        description: 主键ID
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      businessExampleField:
        description: 展示值
        type: string
      mapData:
        additionalProperties:
          type: string
        description: key-val 模式数据
        type: object
    type: object
  business.BasicFile:
    properties:
      CreatedAt:
        description: 创建时间
        type: string
      ID:
        description: 主键ID
        type: integer
      UpdatedAt:
        description: 更新时间
        type: string
      download:
        type: integer
      driver:
        type: integer
      ext:
        type: string
      guid:
        type: string
      mapData:
        additionalProperties:
          type: string
        description: key-val 模式数据
        type: object
      md5:
        type: string
      mediaType:
        type: integer
      module:
        type: integer
      name:
        type: string
      path:
        type: string
      sha1:
        type: string
      size:
        type: integer
      sort:
        type: integer
      usedTime:
        type: integer
      userId:
        type: integer
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /businessExample/createBusinessExample:
    post:
      consumes:
      - application/json
      parameters:
      - description: BusinessExample模型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.BusinessExample'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建BusinessExample
      tags:
      - BusinessExample
  /businessExample/deleteBusinessExample:
    delete:
      consumes:
      - application/json
      parameters:
      - description: BusinessExample模型
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.BusinessExample'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除BusinessExample
      tags:
      - BusinessExample
  /businessExample/findBusinessExample:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - description: 展示值
        in: query
        name: businessExampleField
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询BusinessExample
      tags:
      - BusinessExample
  /businessExample/getBusinessExampleList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - description: 展示值
        in: query
        name: businessExampleField
        type: string
      - description: 排序方式:升序false(默认)|降序true
        in: query
        name: orderDesc
        type: boolean
      - description: Add By ljd 20210708, 统一在这里增加 OrderKey OrderDesc CreatedAtBegin  CreatedAtEnd
        in: query
        name: orderKey
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取BusinessExample列表
      tags:
      - BusinessExample
  /businessExample/updateBusinessExample:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新BusinessExample
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.BusinessExample'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新BusinessExample
      tags:
      - BusinessExample
...
  /basicFile/createBasicFile:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建BasicFile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.BasicFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建BasicFile
      tags:
      - BasicFile
  /basicFile/deleteBasicFile:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除BasicFile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.BasicFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除BasicFile
      tags:
      - BasicFile
  /basicFile/deleteBasicFileByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除BasicFile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除BasicFile
      tags:
      - BasicFile
  /basicFile/findBasicFile:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: download
        type: integer
      - in: query
        name: driver
        type: integer
      - in: query
        name: ext
        type: string
      - in: query
        name: guid
        type: string
      - in: query
        name: md5
        type: string
      - in: query
        name: mediaType
        type: integer
      - in: query
        name: module
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: path
        type: string
      - in: query
        name: sha1
        type: string
      - in: query
        name: size
        type: integer
      - in: query
        name: sort
        type: integer
      - in: query
        name: usedTime
        type: integer
      - in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询BasicFile
      tags:
      - BasicFile
  /basicFile/getBasicFileList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: CreatedAt
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 更新时间
        in: query
        name: UpdatedAt
        type: string
      - in: query
        name: download
        type: integer
      - in: query
        name: driver
        type: integer
      - in: query
        name: ext
        type: string
      - in: query
        name: guid
        type: string
      - in: query
        name: md5
        type: string
      - in: query
        name: mediaType
        type: integer
      - in: query
        name: module
        type: integer
      - in: query
        name: name
        type: string
      - description: 排序方式:升序false(默认)|降序true
        in: query
        name: orderDesc
        type: boolean
      - description: Add By ljd 20210708, 统一在这里增加 OrderKey OrderDesc CreatedAtBegin  CreatedAtEnd
        in: query
        name: orderKey
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: path
        type: string
      - in: query
        name: sha1
        type: string
      - in: query
        name: size
        type: integer
      - in: query
        name: sort
        type: integer
      - in: query
        name: usedTime
        type: integer
      - in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取BasicFile列表
      tags:
      - BasicFile
  /basicFile/quickEdit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 快速更新BasicFile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.BasicFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 快速更新
      tags:
      - QuickEdit
  /basicFile/updateBasicFile:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新BasicFile
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/business.BasicFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新BasicFile
      tags:
      - BasicFile
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"